Move env_replace into envtools
